geoip_lookup/requirements.txt
geoip2
pandas
progressbar2